+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`-` Reviewers can now see both their own score and the median score in the review dashboard.
 - :feature:`-` People without edit or delete permissions are not shown the edit or delete buttons on submissions anymore.
 - :bug:`-` When updating the settings for an event with a custom domain, pretalx would mistakenly show an error message that this domain was already in use for another event.
 - :release:`1.1.2 <2020-01-10>`